Switch pixels to be floating point numbers

+++ src/stage/viewport_rect.rs
@@ -1,8 +1,8 @@
 // A point, measured in pixels, that is relative to the viewport. These would
 // be pixel coordinates as experienced by the user.
 pub struct ViewportRect {
-    pub x: u32,
-    pub y: u32,
-    pub width: u32,
-    pub height: u32,
+    pub x: f32,
+    pub y: f32,
+    pub width: f32,
+    pub height: f32,
 }
